rename datasource to mailer
