+++ basxconnect/mailer_integration/mailchimp/mailer.py
@@ -4,18 +4,15 @@ from typing import List
 import mailchimp_marketing
 from django.conf import settings
 
-from basxconnect.mailer_integration.abstract import abstract_datasource
-from basxconnect.mailer_integration.abstract.abstract_datasource import (
-    MailerPerson,
-    MailingInterest,
-)
+from basxconnect.mailer_integration.abstract import mailer
+from basxconnect.mailer_integration.abstract.mailer import MailerPerson, MailingInterest
 from basxconnect.mailer_integration.mailchimp.parsing import (
     create_mailer_person_from_raw,
 )
 from basxconnect.mailer_integration.models import Interest
 
 
-class MailchimpDatasource(abstract_datasource.Datasource):
+class Mailchimp(mailer.AbstractMailer):
     def __init__(self) -> None:
         self.client = mailchimp_marketing.Client()
         self.client.set_config(
